api/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
langgraph
redis
